add the router in App.js, refactor the welcomescreen
